Move diag.py to the topmost package
